+++ src/client/helpers/index.ts
@@ -1,10 +1,16 @@
-export * from './classes.js';
-export * from './direction.js';
-export * from './isRTL.js';
-export * from './query.js';
-export * from './queryAll.js';
-export * from './slots.js';
-export * from './styles.js';
-export * from './toUnit.js';
-
-export { getConfig, host, isServer, on, off, setConfig } from '../utils/index.js';
+export {
+  classes,
+  direction,
+  getConfig,
+  host,
+  isRTL,
+  isServer,
+  on,
+  off,
+  query,
+  queryAll,
+  slots,
+  styles,
+  toUnit,
+  setConfig
+} from '../utils/index.js';