+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.3] - 2023-05-06
+
+### Added
+
+- Added the `host` element detector.
+
+### Removed
+
+- Removed the `Attribute` decorator.
+
+### Updated
+
+- Updated the internal `utils` path.
+
 ## [0.7.2] - 2023-04-19
 
 ### Updated